Day 7: Did it in std c++ (c++20)
Took about 30 minutes and brute forced in the end. You might be always
able to check the mean? Also the calculation can be much efficient
(certainly in part 1) but it worked.
I did this in a non IDE and it does make me feel about about how I
never know the includes and stuff...