Day 8: Again in c++ as that was nice last time (c++20)
This time I don't see how you would not brute force it without some
transformation to graph or something. Was a little distracted so not
sure how long it took me but once I had the input it was just writing
the checks for all the digits so let's say ~50 minutes or something.
I'm hoping for the day bruteforce doesn't work and I'm stuck and need
to actually be smart.