Day 17: In python took ~40 minutes
Had some confusion about if it could be in the area after every
change or just every step but it turned out I just returned somewhere
where it should have been a check. Thought you might have been able to
do binary search or something but it is not continuous.
Although it might be that there are a number of ranges not sure how
you would be able to limit it to some feasible region.