Day 19: In python took at 2 hours+ since I did not do it in one go
This one was actually tricky with all the rotations
In the end I only used transformations with determinant 1 to make
inverting easy. This came from a post on the subreddit which mentioned
some transformations are mirroring which was not allowed and you can
check it like that. And it happens to be useful for the inverses.

Once part I of the sample worked the rest was easy.
Using the distance to figure out which points are the same was also
taken as inspiration from another solution.

(It turns out the determinant check was not needed as I get correct
results for both parts without it even though I'm not sure if those
inverses are correct. Though this could be just lucky as well.)