Day 25: Python took about 20 minutes
Missed the fact that east went first in the beginning. Fairly easy
otherwise and quite a nice finish.
I definitely had fun but after a while the problems are more annoying to
read with huge blocks of text with one crucial non highlighted detail.
We'll see if I do it again next year. I would like to get a point but with
the starting time and speed of other solvers (I don't think I was ever
close to #100 apart from like 2 days or something).